Made the oversampling available to the chip classification task as well
