+++ rastervision/backend/pytorch_chip_classification_config.py
@@ -24,7 +24,8 @@ class TrainOptions():
                  train_prop=None,
                  train_count=None,
                  log_tensorboard=None,
-                 run_tensorboard=None):
+                 run_tensorboard=None,
+                 oversample=None):
         self.batch_size = batch_size
         self.weight_decay = weight_decay
         self.lr = lr
@@ -72,7 +73,8 @@ class PyTorchChipClassificationConfigBuilder(SimpleBackendConfigBuilder):
                            train_prop=1.0,
                            train_count=None,
                            log_tensorboard=True,
-                           run_tensorboard=True):
+                           run_tensorboard=True,
+                           oversample=None):
         """Set options for training models.
 
         Args:
@@ -112,6 +114,11 @@ class PyTorchChipClassificationConfigBuilder(SimpleBackendConfigBuilder):
             run_tensorboard: (bool) if True, run a Tensorboard server at
                 port 6006 that uses the logs generated by the log_tensorboard
                 option
+            oversample: (dict or None) of form
+                {'rare_class_ids': <list of class ids>, 'rare_target_prop': <float>}
+                This will make it so chips containing any labels in rare_class_ids
+                will be sampled with a probability of rare_target_prop. This is
+                to help cope with severely imbalanced datasets.
         """
         b = deepcopy(self)
         b.train_opts = TrainOptions(
@@ -128,7 +135,8 @@ class PyTorchChipClassificationConfigBuilder(SimpleBackendConfigBuilder):
             train_prop=train_prop,
             train_count=train_count,
             log_tensorboard=log_tensorboard,
-            run_tensorboard=run_tensorboard)
+            run_tensorboard=run_tensorboard,
+            oversample=oversample)
         return b
 
     def with_pretrained_uri(self, pretrained_uri):
